Fix sending signal from CGI
I thought it would make sense to use `true` to restrict the
processes to the ones running on the same server. However,
it doesn’t seem to work as expected, because
buildServerString() returns the fully qualified host name
(such as www.example.com) when running in php-fpm, whereas
it returns the "internal" host name (as returned by
gethostname()) when running from the cake console.
As a result, the CGI code mistakenly assumes the worker
process is running on a different machine and avoids
sending the signal. Let’s make it working out of the box
for users willing to send a signal from CGI to CLI.